refactor: #149 - 링크 삭제 api 함수 파일 이동
