+++ src/app/api/space/[spaceId]/links/route.ts
@@ -28,3 +28,24 @@ export async function GET(
     )
   }
 }
+
+export async function DELETE(req: NextRequest) {
+  const { token } = useServerCookie()
+  const { searchParams } = new URL(req.url)
+  const spaceId = searchParams.get('spaceId')
+  const linkId = searchParams.get('linkId')
+  const path = `/spaces/${spaceId}/links/${linkId}`
+  const headers = {
+    Authorization: `Bearer ${token}`,
+  }
+
+  try {
+    const data = await apiServer.delete(path, {}, headers)
+    return NextResponse.json(data)
+  } catch (error: any) {
+    return NextResponse.json(
+      { error: error.response.data.message },
+      { status: error.response.status },
+    )
+  }
+}
